Further reorganisation and relocation of comments
This commit expands upon the past few by relocating some comments and shifting a few commands to allow readability to be greatly increased. 
The arbitrary "NAME" variable that is only used once now gets a complement: "VERSION" which makes the button def look nicer and make it clear what is happening. 

In my version on fusion i also made one called "CHECKBOX_DESCRIPTION" which contained: "Enables or disables the movement of components by dragging in the canvas." which only further increases the readability but is too pointless to put into a commit